backend/pyproject.toml
[build-system]
requires = [ "poetry-core",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "olx-killer"
version = "0.1.0"
description = ""
authors = [ "Fredy nstanislass@gmail.com",]
readme = "README.md"

[tool.isort]
force_single_line = true
line_length = 120
order_by_type = false
profile = "black"

[tool.black]
line-length = 120
target-version = ['py312']
skip-string-normalization = true
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.poetry.dependencies]
authlib = "^1.3.1"
deep-translator = "^1.11.4"
django = "^5.0.7"
django-anymail = "^11.0.1"
django-cors-headers = "^4.4.0"
django-environ = "^0.11.2"
django-extensions = "^3.2.3"
django-modeltranslation = "^0.19.7"
django-rosetta = "^0.10.0"
djangorestframework = "^3.15.2"
factory-boy = "^3.3.0"
gunicorn = "^23.0.0"
ipython = "^8.26.0"
pillow = "^10.4.0"
polib = "^1.2.0"
psycopg2-binary = "^2.9.9"
python = "^3.12"
python-slugify = "^8.0.4"
requests = "^2.32.3"
unidecode = "^1.3.8"
black = "^24.8.0"

[tool.poetry.dependencies.djangorestframework-simplejwt]
extras = [ "crypto",]
version = "^5.3.1"

[tool.poetry.dependencies.drf-spectacular]
extras = [ "sidecar",]
version = "^0.27.2"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.4.6"
isort = "^5.13.2"
flake8 = "^7.1.0"
